LRC: add skip-calls option to skip first N calls
We need this option in order to test all the calls in the program.
Otherwise a program stops after first failed call which makes it
impossible to test the rest of the calls. This commit adds support for
this option to both randomcrash launcher as command-line argument and
the library, that takes this option from LRC_CONFIG environment
variable.

This commit also replaces val_int of config_opt structure with
val_long in order to store bigger N argument of 'skip-calls'.

Signed-off-by: Vladimir Levijev <vladimir.levijev@gmail.com>